+++ apps/backend/jest.config.ts
@@ -1,0 +1,6 @@
+module.exports = {
+  preset: '../../jest.preset.ts',
+  displayName: 'backend',
+  collectCoverageFrom: ['src/**/*.ts', '!src/**/*.test.ts', '!src/index.ts'],
+  coverageDirectory: './coverage',
+};