+++ apps/backend/src/app/app.module.ts
@@ -3,16 +3,14 @@ import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { APP_GUARD } from '@nestjs/core';
 import { ThrottlerGuard, ThrottlerModule } from '@nestjs/throttler';
-import * as Joi from 'joi';
-import { configuration } from './configuration/configuration';
-import { ConfigurationOptions } from './configuration/configuration.interface';
+import { ConfigurationOptions } from './configuration/configuration.interface.js';
+import { configuration } from './configuration/configuration.js';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       load: [configuration],
       envFilePath: [`.env.${process.env.NODE_ENV}.local`, `.env.${process.env.NODE_ENV}`],
-      validationSchema: Joi.object({ NODE_ENV: Joi.string().valid('development', 'test') }),
       isGlobal: true,
       cache: true,
     }),