+++ apps/backend/src/app/bootstrap/bootstrap.ts
@@ -1,23 +1,21 @@
-/* eslint-disable unicorn/prefer-module */
 import { ApplicationLogger } from '@my-events/nestjs-common';
 import { INestApplication } from '@nestjs/common';
-import { HttpsOptions } from '@nestjs/common/interfaces/external/https-options.interface';
+import { HttpsOptions } from '@nestjs/common/interfaces/external/https-options.interface.js';
 import { ConfigService } from '@nestjs/config';
 import { NestFactory } from '@nestjs/core';
 import { ExpressAdapter } from '@nestjs/platform-express';
-import 'tslib';
-import { AppModule } from '../app.module';
-import { ApplicationOptions, ConfigurationOptions } from '../configuration/configuration.interface';
-import { secureEntrypoint } from '../setup/entrypoint.setup';
-import { configureSwagger } from '../setup/swagger.setup';
+import { AppModule } from '../app.module.js';
+import { ApplicationOptions, ConfigurationOptions } from '../configuration/configuration.interface.js';
+import { secureEntrypoint } from '../setup/entrypoint.setup.js';
+import { configureSwagger } from '../setup/swagger.setup.js';
 
 export const bootstrap = async (): Promise<INestApplication> => {
   // Load HTTPs configuration
-  const enableHTTPs = process.env['APP_TLS_ENABLED'] === 'true';
+  const enableHTTPs = process.env.APP_TLS_ENABLED === 'true';
   const httpsOptions: HttpsOptions = {
-    cert: process.env['APP_TLS_CERTIFICATE']?.replaceAll('\\n', '\n'),
-    key: process.env['APP_TLS_KEY']?.replaceAll('\\n', '\n'),
-    ciphers: process.env['APP_TLS_ALLOWED_CIPHERS'],
+    cert: process.env.APP_TLS_CERTIFICATE?.replaceAll('\\n', '\n'),
+    key: process.env.APP_TLS_KEY?.replaceAll('\\n', '\n'),
+    ciphers: process.env.APP_TLS_ALLOWED_CIPHERS,
     honorCipherOrder: true,
   };