+++ apps/backend/src/app/setup/entrypoint.setup.ts
@@ -1,26 +1,23 @@
-/* eslint-disable unicorn/prefer-module */
 import { ApplicationLogger } from '@my-events/nestjs-common';
 import { HttpStatus, INestApplication } from '@nestjs/common';
 import { NextFunction, Request, Response } from 'express';
 import helmet from 'helmet';
-import 'tslib';
-import { ApplicationOptions } from '../configuration/configuration.interface';
+import { ApplicationOptions } from '../configuration/configuration.interface.js';
 
-export const secureEntrypoint = (application: INestApplication, applicationOptions: ApplicationOptions) => {
-  const applicationlogger = application.get(ApplicationLogger);
+export const secureEntrypoint = (app: INestApplication, applicationOptions: ApplicationOptions) => {
+  const applicationlogger = app.get(ApplicationLogger);
 
   // Configure Express
-  application.setGlobalPrefix(applicationOptions.basePath);
-  application.use(helmet());
-  application.enableCors({ origin: applicationOptions.origin });
-  application.enableVersioning();
+  app.setGlobalPrefix(applicationOptions.basePath);
+  app.use(helmet());
+  app.enableCors({ origin: applicationOptions.origin });
 
   // Log all incoming requests
-  application.use((request: Request, response: Response, next: NextFunction) => {
+  app.use((request: Request, response: Response, next: NextFunction) => {
     applicationlogger.debug(`${request.method} ${request.url}`, 'Express');
     next();
     response.on('finish', () => {
-      const status = response.statusCode;
+      const status = response.statusCode as HttpStatus;
       if (status >= HttpStatus.INTERNAL_SERVER_ERROR) {
         applicationlogger.error(`${request.method} ${request.url} ${response.statusCode}`, 'Express');
       } else if (status >= HttpStatus.BAD_REQUEST) {