+++ apps/backend/src/cli/cli.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import 'tslib';
-import { SwaggerCommand } from './commands/swagger.command';
-
-@Module({
-  providers: [SwaggerCommand],
-})
-export class CliModule {}