+++ apps/backend/src/main.ts
@@ -1,5 +1,5 @@
-/* eslint-disable jest/require-hook */
-import { bootstrap } from './app/bootstrap/bootstrap';
+import { bootstrap } from './app/bootstrap/bootstrap.js';
 
 // Launch the application
+// eslint-disable-next-line @typescript-eslint/no-floating-promises
 bootstrap();