+++ libs/nestjs-common/jest.config.ts
@@ -1,9 +1,6 @@
 module.exports = {
+  preset: '../../jest.preset.ts',
   displayName: 'nestjs-common',
-  preset: '../../jest.preset.js',
-  transform: {
-    '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.json' }],
-  },
   collectCoverageFrom: ['src/**/*.ts', '!src/**/*.test.ts', '!src/index.ts'],
   coverageDirectory: './coverage',
 };