+++ libs/nestjs-common/src/lib/logger/logger-module.definition.ts
@@ -1,5 +1,5 @@
 import { ConfigurableModuleBuilder } from '@nestjs/common';
-import { LoggerModuleOptions } from './logger-module.options';
+import { LoggerModuleOptions } from './logger-module.options.js';
 
 export const { ConfigurableModuleClass, MODULE_OPTIONS_TOKEN } =
   new ConfigurableModuleBuilder<LoggerModuleOptions>().build();