+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.3.0-SNAPSHOT (2024-04-11)
+
+
+### 🚀 Features
+
+- update Husky configuration ([b0fed85](https://github.com/Sitedia/my-events/commit/b0fed85))
+
+### 🩹 Fixes
+
+- update Husky configuration ([20882c3](https://github.com/Sitedia/my-events/commit/20882c3))
+
 ## 0.2.0 (2024-04-06)
 
 ### 🚀 Features